refactor: break down media.tsx into focused modules
Split the 535-line media.tsx file into smaller, focused components:
- base-display.tsx: Core display components (BaseEntityDisplay, PartialMetadataDisplay, MediaScrollArea)
- rating-overlay.tsx: Rating display logic (DisplayAverageRatingOverlay)
- display-items.tsx: Main display item components (MetadataDisplayItem, MetadataGroupDisplayItem, PersonDisplayItem)
- menu-items.tsx: Menu-related components (ToggleMediaMonitorMenuItem, MarkEntityAsPartialMenuItem)

Updated all import statements across 12 files to reference the new modular structure.
All existing functionality is preserved while improving code organization and maintainability.
