+++ apps/frontend/app/routes/_dashboard.media.genre.$id._index.tsx
@@ -16,7 +16,7 @@ import {
 import { useLoaderData } from "react-router";
 import { z } from "zod";
 import { ApplicationGrid } from "~/components/common";
-import { MetadataDisplayItem } from "~/components/media";
+import { MetadataDisplayItem } from "~/components/media/display-items";
 import { pageQueryParam } from "~/lib/shared/constants";
 import { useAppSearchParam } from "~/lib/shared/hooks";
 import {
